+++ sdk/servicebus/service-bus/src/models.ts
@@ -326,3 +326,17 @@ export interface PeekMessagesOptions extends OperationOptionsBase {
    */
   omitMessageBody?: boolean;
 }
+
+/**
+ * Options to configure messages deletion.
+ */
+export interface DeleteMessagesOptions extends OperationOptionsBase {
+  /**
+   * If specified, only messages enqueued before this time are deleted.
+   */
+  beforeEnqueueTimeUtc?: Date;
+  /**
+   * If specified, up to `maxMessageCount` messages will be deleted.
+   */
+  maxMessageCount?: number;
+}